#20 - Добавили поддержку HttpRange в AttachmentEndpoint

+++ dicts/src/test/java/com/backstage/app/dict/service/CommonDictDataServiceTest.java
@@ -996,7 +996,7 @@ public class CommonDictDataServiceTest extends CommonTest
 		var attachment = attachmentService.addAttachment(Objects.requireNonNull(fileResource.getFilename()), MediaType.IMAGE_PNG_VALUE, UserInfo.SYSTEM_USER_ID, bytes);
 		var savedData = attachmentService.getAttachmentData(attachment.getId());
 
-		assertEquals(attachment.getSize(), (Integer) bytes.length);
+		assertEquals(attachment.getSize(), bytes.length);
 		assertArrayEquals(bytes, IOUtils.toByteArray(savedData.getInputStream()));
 
 		return attachment;
